update deps, sveltekit v2 and svelte v5
