--- .eslintrc.cjs
@@ -1,15 +1,0 @@
-module.exports = {
-	root: true,
-	extends: ['eslint:recommended', 'prettier'],
-	plugins: ['svelte3'],
-	overrides: [{ files: ['*.svelte'], processor: 'svelte3/svelte3' }],
-	parserOptions: {
-		sourceType: 'module',
-		ecmaVersion: 2019
-	},
-	env: {
-		browser: true,
-		es2017: true,
-		node: true
-	}
-};
