+++ .nvmrc
@@ -1,0 +1,1 @@
+v22.11.0
