+++ .prettierignore
@@ -1,4 +1,4 @@
-.svelte/**
-static/**
-build/**
-node_modules/**
+# Package Managers
+package-lock.json
+pnpm-lock.yaml
+yarn.lock
