+++ .prettierrc
@@ -3,7 +3,13 @@
 	"singleQuote": true,
 	"trailingComma": "none",
 	"printWidth": 100,
-	"svelteSortOrder": "styles-scripts-markup",
-	"svelteIndentScriptAndStyle": false,
-	"svelteBracketNewLine": false
+	"plugins": ["prettier-plugin-svelte"],
+	"overrides": [
+		{
+			"files": "*.svelte",
+			"options": {
+				"parser": "svelte"
+			}
+		}
+	]
 }
