+++ package.json
@@ -1,38 +1,41 @@
 {
 	"name": "ign-dem-grabber",
-	"version": "0.0.1",
+	"version": "1.0.0",
 	"scripts": {
-		"dev": "svelte-kit dev",
-		"build": "svelte-kit build",
+		"dev": "vite dev",
+		"build": "vite build",
+		"preview": "vite preview",
 		"deploy": "npm run build && gh-pages -t -d build",
-		"start": "svelte-kit start",
-		"lint": "prettier --check . && eslint --ignore-path .gitignore .",
+		"lint": "prettier --check . && eslint .",
 		"format": "prettier --write .",
 		"convert-maps": "bash scripts/convert-maps.sh"
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-static": "^1.0.0-next.4",
-		"@sveltejs/kit": "next",
-		"eslint": "^7.22.0",
-		"eslint-config-prettier": "^8.1.0",
-		"eslint-plugin-svelte3": "^3.1.0",
-		"gh-pages": "^3.1.0",
-		"mapshaper": "^0.5.41",
-		"prettier": "~2.2.1",
-		"prettier-plugin-svelte": "^2.2.0",
-		"shapefile": "^0.6.6",
-		"svelte": "^3.49.0",
-		"topojson-server": "^3.0.1",
-		"topojson-simplify": "^3.0.3",
-		"vite": "^2.1.0"
+		"@sveltejs/adapter-static": "3.0.8",
+		"@sveltejs/kit": "2.15.1",
+		"@sveltejs/vite-plugin-svelte": "4.0.0",
+		"@eslint/compat": "1.2.4",
+		"eslint": "9.17.0",
+		"eslint-config-prettier": "9.1.0",
+		"eslint-plugin-svelte": "2.46.1",
+		"gh-pages": "6.2.0",
+		"mapshaper": "0.6.102",
+		"prettier": "3.4.2",
+		"prettier-plugin-svelte": "3.3.2",
+		"shapefile": "0.6.6",
+		"svelte": "5.16.0",
+		"topojson-server": "3.0.1",
+		"topojson-simplify": "3.0.3",
+		"vite": "5.4.11",
+		"globals": "15.0.0"
 	},
 	"type": "module",
 	"dependencies": {
-		"d3-fetch": "^2.0.0",
-		"d3-geo": "^2.0.1",
-		"d3-geo-projection": "^3.0.0",
-		"d3-selection": "^2.0.0",
-		"maplibre-gl": "^1.14.0",
-		"topojson-client": "^3.1.0"
+		"d3-fetch": "3.0.1",
+		"d3-geo": "3.1.1",
+		"d3-geo-projection": "4.0.0",
+		"d3-selection": "3.0.0",
+		"maplibre-gl": "4.7.1",
+		"topojson-client": "3.1.0"
 	}
 }
