+++ src/app.html
@@ -3,7 +3,7 @@
 	<head>
 		<meta charset="utf-8" />
 		<title>Descarga modelos digitales de elevación</title>
-		<link rel="icon" href="https://martingonzalez.net/ign-dem-grabber/favicon.ico" />
+		<link rel="icon" href="%sveltekit.assets%/favicon.ico" />
 
 		<meta name="title" content="Descarga modelos digitales de elevación">
 		<meta name="description" content="Explora y descarga de forma sencilla los modelos digitales de elevación (DEM) del Instituto Geográfico Nacional en 2m, 5m, 25m y 200m de resolución">
@@ -21,9 +21,9 @@
 		<link rel="preconnect" href="https://fonts.gstatic.com">
 		<link href="https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,300;0,400;0,700;1,300&display=swap" rel="stylesheet">
 		<meta name="viewport" content="width=device-width, initial-scale=1" />
-		%svelte.head%
+		%sveltekit.head%
 	</head>
-	<body>
-		<div id="svelte">%svelte.body%</div>
+	<body data-sveltekit-preload-data="hover">
+		<div style="display: contents">%sveltekit.body%</div>
 	</body>
 </html>
