--- src/routes/$layout.svelte
@@ -1,5 +1,0 @@
-<script>
-	import '../app.css';
-</script>
-
-<slot />
