+++ svelte.config.js
@@ -1,0 +1,24 @@
+import adapter from '@sveltejs/adapter-static';
+// import pkg from './package.json';
+
+/** @type {import('@sveltejs/kit').Config} */
+const config = {
+	kit: {
+		// By default, `npm run build` will create a standard Node app.
+		// You can create optimized builds for different platforms by
+		// specifying a different adapter
+		adapter: adapter(),
+
+		paths: {
+			base: '/ign-dem-grabber'
+		}
+
+		// vite: {
+		// 	ssr: {
+		// 		noExternal: Object.keys(pkg.dependencies || {})
+		// 	}
+		// }
+	}
+};
+
+export default config;
